make tests run (#10)
🐿 v2.8.0
